requirements.txt
matplotlib==3.10
numpy==2.0
pandas==2.3
lightgbm==4.6
xgboost==2.0
scikit-learn==1.7
pyyaml==6.0
catboost
ipython
streamlit==1.4
altair==5.5